# desktop_app/openapi/archestra/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Archestra API",
    "version": "0.0.1"
  },
  "components": {
    "schemas": {
...
      "McpServerConfigInput": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "command"
        ],
        "additionalProperties": false
      },
      "McpServerUserConfigValuesInput": {
        "type": "object",
        "additionalProperties": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "number"
            },
            {
              "type": "boolean"
            },
            {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          ]
        }
      },
...
      "McpServerInstallInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "pattern": "^[A-Za-z0-9-\\s]{1,63}$"
          },
          "serverConfig": {
            "$ref": "#/components/schemas/McpServerConfigInput"
          },
          "userConfigValues": {
            "$ref": "#/components/schemas/McpServerUserConfigValuesInput"
          },
          "oauthProvider": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "google",
                  "slack",
                  "slack-browser",
                  "linkedin-browser"
                ]
              },
              {
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "oauthAccessToken": {
            "type": "string"
          },
          "oauthRefreshToken": {
            "type": "string"
          },
          "oauthExpiryDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          }
        },
        "required": [
          "displayName",
          "serverConfig",
          "oauthProvider"
        ]
      },
...
      "McpServerConfig": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "command"
        ],
        "additionalProperties": false
      },
      "McpServerUserConfigValues": {
        "type": "object",
        "additionalProperties": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "number"
            },
            {
              "type": "boolean"
            },
            {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          ]
        }
      },
      "McpServer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "serverConfig": {
            "$ref": "#/components/schemas/McpServerConfig"
          },
          "userConfigValues": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/McpServerUserConfigValues"
              },
              {
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "oauthAccessToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "oauthRefreshToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "oauthExpiryDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "serverConfig",
          "userConfigValues",
          "oauthAccessToken",
          "oauthRefreshToken",
          "oauthExpiryDate",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/mcp_server/install": {
      "post": {
        "operationId": "installMcpServer",
        "tags": [
          "MCP Server"
        ],
        "description": "Install an MCP server. Either from the catalog, or a customer server",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerInstallInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mcp_server/start_oauth": {
      "post": {
        "operationId": "startMcpServerOauth",
        "tags": [
          "MCP Server",
          "OAuth"
        ],
        "description": "Start MCP server OAuth flow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "catalogName": {
                    "type": "string"
                  },
                  "installData": {
                    "$ref": "#/components/schemas/McpServerInstallInput"
                  }
                },
                "required": [
                  "catalogName",
                  "installData"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authUrl": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "authUrl",
                    "state"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/oauth/callback": {
      "get": {
        "operationId": "oauthCallback",
        "tags": [
          "OAuth"
        ],
        "description": "OAuth callback endpoint for provider redirects",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "code",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "state",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "error",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "error_description",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/mcp_server/complete_oauth": {
      "post": {
        "operationId": "completeMcpServerOauth",
        "tags": [
          "MCP Server",
          "OAuth"
        ],
        "description": "Complete MCP server OAuth flow and install with tokens",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "service": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "access_token": {
                    "type": "string"
                  },
                  "refresh_token": {
                    "type": "string"
                  },
                  "expiry_date": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "required": [
                  "service",
                  "state"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}